Comments. Moved merge class
