+++ visualizer_main.py
@@ -12,6 +12,7 @@ clock = pygame.time.Clock()
 pygame.init()
 
 class Visualizer():
+    "main visualizer state, manager of states"
     font_squid = pygame.font.Font(os.path.join(config.fonts_dir, "Game Of Squids.ttf"), 20)
     font_sorting_overlay_dir = os.path.join(config.fonts_dir, "Cotton Butter.ttf")
     def __init__(self) -> None:
@@ -30,8 +31,8 @@ class Visualizer():
         self.load_states()
         self.delay_input = 0
 
-    # main game loop
     def visualizer_loop(self, run_once=False):
+        """main game loop"""
         while self.playing:
             #clock.tick(60)  # untick to not burn cpu
             if self.resize_delay < pygame.time.get_ticks():
@@ -42,8 +43,8 @@ class Visualizer():
                 if run_once:
                     return
 
-    #  check events and set actions dicts with resaults to check from other states
     def check_events(self):
+        """check events and set actions dicts with resaults to check from other states"""
         for event in pygame.event.get():
             if event.type == pygame.VIDEORESIZE:  # refresh, resize everything and render the entire screen on resize
                 self.display_reset = True
@@ -94,24 +95,26 @@ class Visualizer():
                 if event.button == 1:
                     self.actions["left_mouse"] = False
 
-    # update top state in stack
     def update(self):
+        """Update the top state of the stack"""
         self.state_stack[-1].update(self.dt, self.actions)
 
     def render(self):
+        """render the entire screen or bar area"""
         self.state_stack[-1].render(self.WINDOW)
         if self.display_reset:  # check if entire display needs to be rendered due to changes
             pygame.display.flip()
             self.display_reset = False
-        pygame.display.flip()
+        pygame.display.update(self.SCREEN_WIDTH *0, self.SCREEN_HEIGHT*0.4, self.SCREEN_WIDTH *1, self.SCREEN_HEIGHT*0.6)
 
     def get_dt(self):
+        """get delta time"""
         now = time.time()
         self.dt = now - self.prev_time
         self.prev_time = now
 
-    #  helper function for displaying text
     def draw_text(self, surface, text, color, x, y, font=None, scale=False):
+        """helper function for displaying text"""
         if not font or font == "font_squid":
             font = Visualizer.font_squid
         elif font == "font_sorting_overlay":
@@ -126,16 +129,19 @@ class Visualizer():
         return text_rect
 
     def load_assets(self):
+        """load assests on startup"""
         ft = pygame.font.Font
         self.font_table = {}
         for i in range(0, 100):  # create different font size for responsiveness
             self.font_table[i] = ft(self.font_sorting_overlay_dir, i)
 
     def load_states(self):
+        """Load and stack MainMenu state"""
         self.main_menu_screen = MainMenu(self)
         self.state_stack.append(self.main_menu_screen)
 
     def reset_keys(self):
+        """reset_inputs"""
         for action in self.actions:
             self.actions[action] = False
